Update react-re-render-3_en.md
Since the execution starts from App component, the useEffect will get called in the first render only and state value changes to 1, then "App" will get printed in the console. Then the code moves to the return statement, component A gets called "A" is printed and it's child component is returned so B component is called and "B" is printed next, since C component is returned "C" is printed next and since C returns null, it comes out of the loop and A component is closed. next it goes to D component and "D" gets printed and as it returns null D is closed.